feature(web): extract platform I/O boundaries in engine
