+++ engine/src/core/mod.rs
@@ -1,3 +1,6 @@
+pub mod export_native;
+pub mod import_native;
 pub mod receive;
 pub mod send;
+pub mod storage;
 pub mod types;
